Create c
This lecture is built around short programs that use 'if', 'while', and 'for' statements to address important computational tasks.   Examples included sorting, computing the square root, factoring, and simulating a random process. The lecture concluded with a detailed example illustrating the process of debugging a program. The assignments were great for refreshing my existing understanding of looping structures in Java.