Create code
This lecture described Java's built-in array data structure that supports computation with a large sequence of values of the same type, with several examples, including shuffling a deck of cards, the coupon collector test for randomness, and random walks in a grid. 
The assignment was delightfully challenging, with the resulting code being useful for real-world applications.