Create code
The lecture described several mechanisms for taking and displaying information through a program, in the form of: text, drawings, and animation. Detailed examples include fractal drawings that model natural phenomena and an animation of a ball bouncing around in the display window.
The assignment was very insightful in using Java for broad spectrum data visualization.